Some more refactoring packages
