+++ staramr/databases/drug/ARGDrugTable.py
@@ -3,7 +3,7 @@ from os import path
 
 class ARGDrugTable:
 
-    DEFAULT_DATA_DIR=path.join(path.dirname(__file__), '..', 'data')
+    DEFAULT_DATA_DIR=path.join(path.dirname(__file__), 'data')
 
     def __init__(self, file):
         self.file = file
